requirements.txt
openai
python-dotenv
# tabulate

# for Gemini proxy
google-generativeai
fastapi
uvicorn
protobuf==3.20.*